requirements.txt
PyYAML
jinja2